changelog
- update Docs
- use url-join
- move decorators to decorator/
- rename decorators
- refactor Prefix decorator
- calc Pager until req.pager getter
