+++ src/decorator/index.ts
@@ -1,0 +1,3 @@
+export { Config, FinalClass, Prefix } from './config';
+
+export { Endpoint, ExpressRouter } from './expressRouter';
