+++ src/index.ts
@@ -4,14 +4,14 @@ import * as joi from 'joi';
 // foundation libraries
 export { express, joi };
 
-// expressApp Interface
+// main lib
 export { ExpressApplication } from './class/expressApplication';
 
 // Middleware
 export { jwtAuth } from './middleware/jwtAuth';
 
 // decorators
-export { expressRouter, endpoint, config, final, prefix } from './decorators';
+export { ExpressRouter, Endpoint, Config, FinalClass, Prefix } from './decorator';
 
 // Interfaces
 export {
