+++ src/middleware/publishPager.ts
@@ -7,22 +7,23 @@ import { Pager } from '../entity/pager';
 
 export function publishPager(getConfig: Function) {
 	return function pager(req: IRequest, res: IResponse, next: NextFunction) {
-		const page = req.query.page || req.body.page || req.headers['page'] || getConfig('defaultPage', PAGE);
-		const per_page = req.query.per_page || req.body.per_page || req.headers['per_page'] || getConfig('defaultPerPage', PER_PAGE);
-		const validationOpts: joi.ValidationOptions = {
-			abortEarly: false,
-			convert: true,
-		};
-		const { error, value } = joi.validate<Pager>({ page, per_page }, pagerValidator(PAGE, PER_PAGE), validationOpts);
+		Object.defineProperty(req, 'pager', {
+			get: (): Pager => {
+				const page = req.query.page || req.body.page || req.headers['page'] || getConfig('defaultPage', PAGE);
+				const per_page = req.query.per_page || req.body.per_page || req.headers['per_page'] || getConfig('defaultPerPage', PER_PAGE);
+				const validationOpts: joi.ValidationOptions = { abortEarly: false, convert: true, };
+				const { error, value } = joi.validate<Pager>({ page, per_page }, pagerValidator(PAGE, PER_PAGE), validationOpts);
 
-		if (error) {
-			const dto = { type: 400, message: error.message, errors: error.details };
+				if (error) {
+					const dto = { type: 400, message: error.message, errors: error.details };
 
-			return res.errorJson(dto);
-		} else {
-			req.pager = new Pager(value);
+					throw res.errorJson(dto);
+				} else {
+					return new Pager(value);
+				}
+			},
+		});
 
-			return next();
-		}
+		return next();
 	};
 }
